+++ packages/admin/src/main.ts
@@ -2,13 +2,13 @@
  * Created by bangbang93 on 2017/8/25.
  */
 'use strict'
+import {createApp} from 'vue'
 import Element from 'element-plus'
 import 'element-plus/dist/index.css'
 import 'font-awesome/scss/font-awesome.scss'
-import {createApp} from 'vue'
 import VueFetch from 'vue-fetch'
-import App from '../pages/admin/index.vue'
-import router from '../router/admin'
+import App from './pages/index.vue'
+import router from './router'
 
 export const app = createApp(App)
