+++ packages/admin/vite.config.ts
@@ -1,0 +1,15 @@
+import vuePlugin from '@vitejs/plugin-vue'
+import {fileURLToPath, URL} from 'node:url'
+import {defineConfig} from 'vite'
+
+// https://vitejs.dev/config/
+export default defineConfig({
+  plugins: [
+    vuePlugin(),
+  ],
+  resolve: {
+    alias: {
+      '@': fileURLToPath(new URL('./src', import.meta.url)),
+    },
+  },
+})
